+++ common/src/main/java/net/mehvahdjukaar/polytone/dimension/DimensionEffectsManager.java
@@ -15,9 +15,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.color.block.BlockColor;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Registry;
 import net.minecraft.core.RegistryAccess;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.RegistryOps;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.CubicSampler;
@@ -36,9 +34,12 @@ import java.util.Set;
 
 public class DimensionEffectsManager extends JsonImgPartialReloader {
 
-    private final Map<ResourceLocation, DimensionEffectsModifier> effectsToApply = new HashMap<>();
+    //we cant store dimensions here since the dimension registry isnt synced to the client
+    //!!map of dimension modifier IDs to modifiers
+    private final MapRegistry< DimensionEffectsModifier> dimensionEffects = new MapRegistry<>("Dimension Effects Modifiers");
 
-    private final Map<ResourceLocation, DimensionEffectsModifier> vanillaEffects = new HashMap<>();
+    //map of dimension ID to modifier
+    private final Map<ResourceLocation, DimensionEffectsModifier> alteredVanillaEffects = new HashMap<>();
 
     private final Object2ObjectMap<DimensionType, IColorGetter> fogColormaps = new Object2ObjectArrayMap<>();
     private final Object2ObjectMap<DimensionType, IColorGetter> terrainFogColormaps = new Object2ObjectArrayMap<>();
@@ -62,7 +63,7 @@ public class DimensionEffectsManager extends JsonImgPartialReloader {
 
         //Dimensions are NOT reloaded with world load. we need to reset vanilla stuff once we have a level
         //whatever happens, we always clear stuff to apply
-        effectsToApply.clear();
+        dimensionEffects.clear();
         fogColormaps.clear();
         skyColormaps.clear();
         sunsetColormaps.clear();
@@ -153,9 +154,7 @@ public class DimensionEffectsManager extends JsonImgPartialReloader {
     }
 
     private void addModifier(ResourceLocation fileId, DimensionEffectsModifier mod, RegistryAccess registryAccess) {
-        for (var h : mod.targets().getTargets(fileId, registryAccess)) {
-            effectsToApply.merge(h.unwrapKey().get().location(), mod, DimensionEffectsModifier::merge);
-        }
+        dimensionEffects.register(fileId, mod);
     }
 
     @Override
@@ -163,43 +162,54 @@ public class DimensionEffectsManager extends JsonImgPartialReloader {
         if (!isLogIn && !needsDynamicApplication) return;
         needsDynamicApplication = false;
 
-        for (var v : vanillaEffects.entrySet()) {
+        //reset vanilla
+        for (var v : alteredVanillaEffects.entrySet()) {
             v.getValue().applyInplace(v.getKey());
         }
 
-        Registry<DimensionType> dimReg = access.registryOrThrow(Registries.DIMENSION_TYPE);
+        //apply to current dimension
+        Level level = Minecraft.getInstance().level;
+        if (level != null) {
+            onDimensionChanged(level, access);
+        }
+    }
 
-        for (var v : effectsToApply.entrySet()) {
-            ResourceLocation dimensionId = v.getKey();
+    public void onDimensionChanged(Level level, RegistryAccess access) {
+        DimensionType currentDim = level.dimensionType();
+        var currentDimHolder = level.dimensionTypeRegistration();
+        ResourceLocation currentDimId = currentDimHolder.unwrapKey().get().location();
+
+        for (var v : dimensionEffects.getEntries()) {
+            ResourceLocation modId = v.getKey();
             DimensionEffectsModifier modifier = v.getValue();
-            var old = modifier.applyInplace(dimensionId);
+            var targets = modifier.targets().compute(modId, access);
+            if (!targets.contains(currentDimHolder)) continue;
 
-            vanillaEffects.put(dimensionId, old);
+            var old = modifier.applyInplace(currentDimId);
 
-            DimensionType dim = dimReg.get(dimensionId);
+            alteredVanillaEffects.put(currentDimId, old);
+
             if (modifier.getFogColormap() instanceof IColorGetter c) {
-                fogColormaps.put(dim, c);
+                fogColormaps.put(currentDim, c);
             }
             if (modifier.getSkyColormap() instanceof Colormap c) {
-                skyColormaps.put(dim, c);
+                skyColormaps.put(currentDim, c);
             }
             if (modifier.getSunsetColormap() instanceof IColorGetter c) {
-                sunsetColormaps.put(dim, c);
+                sunsetColormaps.put(currentDim, c);
             }
             if (modifier.noWeatherFogDarken()) {
-                cancelFogWeatherDarken.put(dim, true);
+                cancelFogWeatherDarken.put(currentDim, true);
             }
-            if (modifier.noWeatherSkyDarken()){
-                cancelSkyWeatherDarken.put(dim, true);
+            if (modifier.noWeatherSkyDarken()) {
+                cancelSkyWeatherDarken.put(currentDim, true);
             }
             if (modifier.cloudLevel().isPresent() && modifier.cloudLevel().get().right().isPresent()) {
-                cloudFunctions.put(dim, modifier.cloudLevel().get().right().get());
+                cloudFunctions.put(currentDim, modifier.cloudLevel().get().right().get());
             }
+            Polytone.LOGGER.info("Applied Custom Dimension Effects Modifier '{}' to dimension '{}'", modId, currentDimHolder);
         }
-        if (!vanillaEffects.isEmpty())
-            Polytone.LOGGER.info("Applied {} Custom Dimension Effects Properties", vanillaEffects.size());
         //we don't clear effects to apply because we need to re apply on world reload
-
     }
 
     @Nullable
@@ -220,8 +230,8 @@ public class DimensionEffectsManager extends JsonImgPartialReloader {
     public void modifyFogMagicNumber(float renderDistanceChunks, LocalFloatRef distance) {
         //no more random sky seam!
         float c = 0.25f;
-        float b = c + (1-c) * renderDistanceChunks / 32.0F;
-        b = 1.0F - (float)Math.pow(b, 0.25);
+        float b = c + (1 - c) * renderDistanceChunks / 32.0F;
+        b = 1.0F - (float) Math.pow(b, 0.25);
         float a = 1 * renderDistanceChunks / 32.0F;
         a = 1.0F - (float) Math.pow(a, 0.25);
         distance.set(b);
@@ -269,7 +279,7 @@ public class DimensionEffectsManager extends JsonImgPartialReloader {
 
     private static float[] lastSunset = null;
 
-    public float @Nullable [] modifySunsetColor(float [] old) {
+    public float @Nullable [] modifySunsetColor(float[] old) {
         IColorGetter colormap = this.sunsetColormaps.get(Minecraft.getInstance().level.dimensionType());
         if (colormap == null) return null;
         var color = colormap.sampleColor(null, ClientFrameTicker.getCameraPos(),