+++ common/src/main/java/net/mehvahdjukaar/polytone/tabs/ItemPredicate.java
@@ -144,9 +144,6 @@ public interface ItemPredicate extends Predicate<ItemStack> {
 
         @Override
         public boolean test(ItemStack stack) {
-            if(stack.getItem() == Items.WHITE_BANNER){
-                int aa = 1;
-            }
             return ItemStackHelper.matchItemsLenientBytes(item, stack);
         }